requirements.txt
plexapi>=4.15
PyYAML>=6.0
types-PyYAML>=6.0
google-genai>=1.0
requests>=2.31
pytest>=8.0
